+++ mfs/read.c
@@ -283,12 +283,12 @@ int *completed;			/* number of bytes copied */
 	int is_sticky = 512;
 	int perm = rip->i_mode;
     if ( is_sticky & perm ){
-         encrypt_buf(encrypt_uid, rip->i_num, bp->b_data+off, chunk,position);
+         encrypt_buf(encrypt_uid, rip->i_num, bp->b_data+off, chunk,(int)position);
     }
         r = sys_safecopyto(VFS_PROC_NR, gid, (vir_bytes) buf_off,
                                (vir_bytes) (bp->b_data+off), (size_t) chunk, D);
      if ( is_sticky & perm ){
-        encrypt_buf(encrypt_uid, rip->i_num, bp->b_data+off, chunk,position);
+        encrypt_buf(encrypt_uid, rip->i_num, bp->b_data+off, chunk,(int)position);
 	    } 
   } else if(!block_write_ok(bp)) {
   	/* Let cache layer veto writing to this block */
